* fix bugs in logging with BIG DATASETS

+++ MyNN/MyNN.MLP.Classic/Backpropagation/EpocheTrainer/Classic/OpenCL/GPU/GPUEpocheTrainer.cs
@@ -403,7 +403,7 @@ namespace MyNN.MLP.Classic.Backpropagation.EpocheTrainer.Classic.OpenCL.GPU
                             {
                                 ConsoleAmbientContext.Console.Write(
                                     "Epoche progress: {0}%, {1}      ",
-                                    (currentIndex*100/data.Count),
+                                    ((long)currentIndex * 100 / data.Count),
                                     DateTime.Now.ToString());
 
                                 ConsoleAmbientContext.Console.ReturnCarriage();
