+++ MyNN/MyNN.MLP.Classic/Backpropagation/EpocheTrainer/TransposedClassic/OpenCL/CPU/CPUTransposeEpocheTrainer.cs
@@ -395,7 +395,7 @@ namespace MyNN.MLP.Classic.Backpropagation.EpocheTrainer.TransposedClassic.OpenC
                             {
                                 ConsoleAmbientContext.Console.Write(
                                     "Epoche progress: {0}%, {1}      ",
-                                    (currentIndex*100/data.Count),
+                                    ((long)currentIndex * 100 / data.Count),
                                     DateTime.Now.ToString());
 
                                 ConsoleAmbientContext.Console.ReturnCarriage();
