+++ MyNN/MyNN.MLP.Dropout/Backpropagation/EpocheTrainer/Dropout/OpenCL/CPU/CPUDropoutEpocheTrainer.cs
@@ -442,7 +442,7 @@ namespace MyNN.MLP.Dropout.Backpropagation.EpocheTrainer.Dropout.OpenCL.CPU
                             {
                                 ConsoleAmbientContext.Console.Write(
                                     "Epoche progress: {0}%, {1}      ",
-                                    (currentIndex*100/data.Count),
+                                    ((long)currentIndex * 100 / data.Count),
                                     DateTime.Now.ToString());
 
                                 ConsoleAmbientContext.Console.ReturnCarriage();
