+++ MyNN/MyNN.MLP.NLNCA/Backpropagation/EpocheTrainer/NLNCA/ClassificationMLP/OpenCL/CPU/CPUNLNCAEpocheTrainer.cs
@@ -447,7 +447,7 @@ namespace MyNN.MLP.NLNCA.Backpropagation.EpocheTrainer.NLNCA.ClassificationMLP.O
                             {
                                 ConsoleAmbientContext.Console.Write(
                                     "Epoche progress: {0}%, {1}      ",
-                                    (currentIndex*100/data.Count),
+                                    ((long)currentIndex * 100 / data.Count),
                                     DateTime.Now.ToString());
 
                                 ConsoleAmbientContext.Console.ReturnCarriage();
